Set the from address and the mail command (#268)
* Set the from address and the mail command

* adjust the domain guessing

* use the already setted variables also for smtps

* use sed to encode characters not correctly encoded by jq

* Use bash function to encode username and password

* simplify logic

* Don't send duplicate Date: header

Asterisk writes the "Date:" header into email bdy, but s-nail writes it again. It isn't possible to disable it from configuration in voicemail.conf or s-nail.rc, therefore here there is a simple bash script that removes the Date: header then pipes the mail to s-nail. It is possible to improve the script further in future, for instance by sending the voicemail to other channels or do STT on the attachment.

* Make it work if SMTP configuration is "use local mail app"